TST: move IPKernelClient class to 'common.py'
